+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # change log for utils
 
+## 2.0.1 2024-05-12 bugfix
+
+- Fixed the runagain feature in testrun 
+- Fixed a localization problem in testcases of "sun"
+
 ## 2.0.0 2023-04-19 release
 
 - Create a full collection of utilities
@@ -8,6 +13,7 @@
 - Refactored all code
 - Refactored all tests
 - Convert it to typescript
+- Added the unittest package
 
 ## 1.2.0 2022-01-03 update